separated notebook function into notebook
