+++ gui/__init__.py
@@ -4,3 +4,4 @@ from . import graph
 from . import mainGui
 from . import registration
 from . import search
+from . import mainNotebook
