+++ gui/notebook/__init__.py
@@ -1,0 +1,3 @@
+from . import titlePanel
+from . import transactionPanel
+from . import yearPanel
